Change Daily page to use new CollapsibleNotesList
