+++ fe/src/components/notes/NotesPage.tsx
@@ -1,18 +1,10 @@
 import React, { useEffect, useState } from "react";
 import "./NotesPage.scss";
-import TextBox from "../textbox/TextBox";
-import {
-  Note,
-  getNotes,
-  removeNote,
-  updateNote,
-  createNote,
-  GetNotesParams,
-} from "./NotesApi";
+import { Note, getNotes, GetNotesParams } from "./NotesApi";
 import { Header } from "../header/Header";
-import { ListWidget, ListWidgetProps } from "../widgets/ListWidget/ListWidget";
 import moment from "moment";
 import { objectIdFromDate } from "../../utils/ObjectId";
+import { CollapsibleNotesList } from "./CollapsibleNotesList/CollapsibleNotesList";
 
 export type NotesPageProps = {
   daily?: boolean;
@@ -21,6 +13,7 @@ export type NotesPageProps = {
 export default function NotesPage(props: NotesPageProps) {
   const [date, setDate] = useState<Date>(moment().startOf("day").toDate());
   const [notes, setNotes] = useState<Note[]>([]);
+  const [lastUpdate, setLastUpdate] = useState("");
 
   let title = "Archive";
   if (props.daily) {
@@ -38,39 +31,9 @@ export default function NotesPage(props: NotesPageProps) {
 
     getNotes(getNotesParams).then((notes) => {
       setNotes(notes);
+      setLastUpdate(moment().format());
     });
   }, [date, props.daily]);
-
-  const handleSubmit = async (note: Partial<Note>) => {
-    const newNote = await createNote(note);
-    setNotes([newNote, ...notes]);
-  };
-
-  const handleItemChanged: ListWidgetProps["onItemChanged"] = (item, index) => {
-    updateNote(item);
-
-    const notesCopy = Array.from(notes);
-    const oldNote = notes[index];
-    let newNote: Note = {
-      ...oldNote,
-      ...item,
-    };
-    let itemsToReinsert = [newNote];
-    if (props.daily && item.archived) {
-      itemsToReinsert = [];
-    }
-
-    notesCopy.splice(index, 1, ...itemsToReinsert);
-    setNotes(notesCopy);
-  };
-
-  const handleItemDeleted = (itemId?: string) => {
-    removeNote(itemId);
-    const notesCopy = Array.from(notes);
-    const index = notesCopy.findIndex((item) => item._id === itemId);
-    notesCopy.splice(index, 1);
-    setNotes(notesCopy);
-  };
 
   const handleArrowClicked = (direction: string) => {
     if (direction === "left") {
@@ -87,14 +50,7 @@ export default function NotesPage(props: NotesPageProps) {
         showArrows={props.daily}
         onArrowClicked={handleArrowClicked}
       />
-      <ListWidget
-        items={notes}
-        onItemChanged={handleItemChanged}
-        onItemDeleted={handleItemDeleted}
-        actions={props.daily ? ["archive"] : ["delete"]}
-        showArchived={props.daily ? false : true}
-      />
-      <TextBox onSubmit={handleSubmit} />
+      <CollapsibleNotesList key={lastUpdate} notes={notes} />
     </div>
   );
 }
